pcli: move command bodies into submodules
This code organization could certainly be improved further, but it means that
not all of the logic is in one main function.
